Multidimensional arrays
Implementing this involved a lot of backtracking and fleshing out
ideas half-implemented before, like the _as_param_ method thing.
All types including arrays must now have an _as_param_ method,
an _update_ method, and a _typecode_. In Ctypes types these are
woven in with FETCH and STORE, but using this standard we *should*
be able to work with user-defined types in calling Functions, etc.

Full write up to follow: API description in docs, technical detail
in blog post.

	modified:   Ctypes.xs
	modified:   lib/Ctypes/Type.pm
	modified:   lib/Ctypes/Type/Array.pm
	modified:   t/Array.t
	modified:   t/types.t