Pointer objects
See t/Pointers.t for usage.

In general, '$ptr' for methods on the object and '$$ptr' for using
the pointer behaviour, e.g.

$$ptr[0] = 5;

The notable exception is that pointer arithmetic is done to the
object:

$ptr++;
$ptr->offset(6);

If you do $$ptr++ Ctypes attempts to construct a new pointer to
the numeric representation of the address of the object (I think);
there's a decent error message for this.

	modified:   MANIFEST
	modified:   lib/Ctypes.pm
	modified:   lib/Ctypes/Type.pm
	modified:   lib/Ctypes/Type/Array.pm
	new file:   lib/Ctypes/Type/Pointer.pm
	modified:   t/Array.t
	new file:   t/Pointer.t