Fixed multiple indirection on Pointers / Arrays
At the minute, the object that was finally passed must either be
explicitly ->_update_'d, or must be the first used after the
call returns (in which case it updates itself automatically).

This should be made automatic using paramflags.

Alternatively, _as_param_ could take an argument of a reference
to the requesting object, or the place the data will be stored and
worked upon. Then data could be 'chased up' the object chain to
ensure freshness.
Interface should probably be through _update_, with objects free
to use their own internal mechanism to know whether their own
data is fresh (i.e. has it been stored again since it was given
out?).

	modified:   lib/Ctypes/Type.pm
	modified:   lib/Ctypes/Type/Array.pm
	modified:   lib/Ctypes/Type/Pointer.pm
	modified:   t/Array.t
	modified:   t/Pointer.t