Simple.t passing many more tests than before
Fixed bug introduced in chages to C:T:Simple.pm in commit 07bf5c708
(First version of class refactor). Misleading names for internally
used properties were probably responsible for the confusion.

Changes to _hook_fetch due to the fix have been documented in POD.

Key is in keeping the value as a property of $object->{_rawvalue}
and not in $object->{_value} directly. $object->{_value} can't be
trusted since $$x = <anything> will squash it in the course of
scalar overload. $object->{_rawvalue}->{VALUE} or ->[1]
is the only location which is checked on input and output.
In effect, $object->{_value} is just a placeholder which is
accessed to allow tie() to do its magic; it shouldn't be used
directly.

Of course, the way everything /should/ be done is to not
cache /anything/ Perl-side, and keep all values as native C
variables / data structures only. This would remove so much
messing about keeping track of values in multiple locations,
cascading changes around objects etc.

Python Ctypes keeps all values as native C. The only reason
I didn't do it that way to begin with is because I was less
comfortable in C and thought it would be too hard. In
retrospect, it would have saved a ton of work.

	modified:   lib/Ctypes/Type/Simple.pm
	modified:   t/Simple.t